Refactor and rename experiment-related files
- Renamed Container.py to data_container.py
- Moved Experiment class to a new location
- Updated import paths in job.py and test files accordingly
