+++ src/easydiffraction/job/experiment/data_container.py
@@ -4,9 +4,8 @@
 
 from easyscience.Objects.core import ComponentSerializer
 
-from easydiffraction.Profiles.Experiment import Experiment
+from easydiffraction.job.experiment.experiment import Experiment
 from easydiffraction.Profiles.Sample import Sample
-
 
 class DataContainer(ComponentSerializer):
     def __init__(self, sim_store: Sample, exp_store: Experiment):
