+++ tests/unit_tests/test_Job.py
@@ -7,7 +7,7 @@ import easydiffraction as ed
 from easydiffraction.calculators.wrapper_factory import WrapperFactory
 from easydiffraction.job.job import DiffractionJob as Job
 from easydiffraction.job.analysis.analysis import Analysis
-from easydiffraction.Profiles.Experiment import Experiment
+from easydiffraction.job.experiment.experiment import Experiment
 from easydiffraction.Profiles.JobType import JobType
 from easydiffraction.Profiles.P1D import Instrument1DCWParameters
 from easydiffraction.Profiles.P1D import Instrument1DTOFParameters
